Java docs ScoreController
